Break up bit_stream into multiple files
