+++ fly/coders/huffman/huffman_decoder.cpp
@@ -2,7 +2,7 @@
 
 #include "fly/literals.h"
 #include "fly/logger/logger.h"
-#include "fly/types/bit_stream/bit_stream.h"
+#include "fly/types/bit_stream/bit_stream_reader.h"
 
 #include <vector>
