+++ fly/coders/huffman/huffman_encoder.cpp
@@ -3,7 +3,7 @@
 #include "fly/coders/huffman/huffman_config.h"
 #include "fly/literals.h"
 #include "fly/logger/logger.h"
-#include "fly/types/bit_stream/bit_stream.h"
+#include "fly/types/bit_stream/bit_stream_writer.h"
 
 #include <algorithm>
 #include <limits>
