+++ fly/types/bit_stream/bit_stream.h
@@ -1,17 +1,12 @@
 #pragma once
 
+#include "fly/types/bit_stream/bit_stream_traits.h"
+#include "fly/types/bit_stream/bit_stream_types.h"
 #include "fly/types/bit_stream/endian.h"
 
-#include <cstdint>
-#include <istream>
 #include <limits>
-#include <type_traits>
 
 namespace fly {
-
-typedef std::uint8_t byte_type;
-typedef std::uint16_t word_type;
-typedef std::uint64_t buffer_type;
 
 /**
  * Base class for writing to and reading from a binary stream. The first byte of
@@ -68,246 +63,6 @@ protected:
     buffer_type m_buffer;
 };
 
-/**
- * Implementation of the BitStream interface for writing to a binary stream.
- *
- * Bits are written to an in-memory byte buffer until that buffer is full, at
- * which point that buffer is flushed to the stream. When done writing, callers
- * should invoke the Finish() method to flush the BitStream header and any bytes
- * remaining in the buffer.
- *
- * @author Timothy Flynn (trflynn89@gmail.com)
- * @version July 7, 2019
- */
-class BitStreamWriter : public BitStream
-{
-public:
-    /**
-     * Constructor. Write the header byte onto the stream.
-     *
-     * The wrapped stream must be both an input and output stream. This allows
-     * going back and rewriting the header byte to contain the number of
-     * zero-filled bits.
-     *
-     * @param iostream The stream to write binary data into.
-     */
-    BitStreamWriter(std::iostream &) noexcept;
-
-    /**
-     * Write a multibyte word to the byte buffer. Flush the buffer to the stream
-     * if it is filled during this operation.
-     *
-     * @param word_type The word to write.
-     */
-    void WriteWord(word_type) noexcept;
-
-    /**
-     * Write a full byte to the byte buffer. Flush the buffer to the stream if
-     * it is filled during this operation.
-     *
-     * @param byte_type The byte to write.
-     */
-    void WriteByte(byte_type) noexcept;
-
-    /**
-     * Write a number of bits to the byte buffer. The least-significant bits in
-     * the provided data type will be written, starting from the position
-     * pointed to by the provided number of bits. Flush the buffer to the stream
-     * if it is filled during this operation.
-     *
-     * @tparam DataType The data type storing the bits to write.
-     *
-     * @param DataType The bits to write.
-     * @param byte_type The number of bits to write.
-     */
-    template <typename DataType>
-    void WriteBits(DataType, byte_type) noexcept;
-
-    /**
-     * If needed, zero-fill the byte buffer, flush it to the stream, and update
-     * the header byte.
-     *
-     * @return bool True if the stream remains in a good state.
-     */
-    bool Finish() noexcept;
-
-private:
-    /**
-     * Flush the header byte onto the stream.
-     *
-     * @param byte_type The number of zero-filled bits in the byte buffer.
-     */
-    void flushHeader(byte_type) noexcept;
-
-    /**
-     * Flush the byte buffer onto the stream.
-     */
-    void flushBuffer() noexcept;
-
-    /**
-     * Flush a byte buffer to the stream.
-     *
-     * @tparam DataType The type of the byte buffer to flush.
-     *
-     * @param DataType The byte buffer to flush.
-     * @param byte_type The number of bytes to flush.
-     */
-    template <typename DataType>
-    void flush(const DataType &, byte_type) noexcept;
-
-    std::iostream &m_stream;
-};
-
-/**
- * Implementation of the BitStream interface for reading from a binary stream.
- *
- * The stream is read in a lazy manner; bytes are not read from the stream until
- * they are needed. The number of bytes read from the stream at once is defined
- * by the size of buffer_type. That buffer is stored in-memory until it has been
- * entirely consumed by the caller, at which point it is refilled.
- *
- * @author Timothy Flynn (trflynn89@gmail.com)
- * @version July 7, 2019
- */
-class BitStreamReader : public BitStream
-{
-public:
-    /**
-     * Constructor. Decode the header byte from the stream. If the header byte
-     * is invalid, the stream's fail bit is set.
-     *
-     * @param iostream The stream to read binary data from.
-     */
-    BitStreamReader(std::istream &) noexcept;
-
-    /**
-     * Read a multibyte word from the byte buffer. Fill the buffer from the
-     * stream if it is fully consumed during this operation.
-     *
-     * @param word_type The location to store the read word.
-     *
-     * @return bool True if the word was successfully read and filling the byte
-     *              buffer was successful (if needed).
-     */
-    bool ReadWord(word_type &) noexcept;
-
-    /**
-     * Read a full byte from the byte buffer. Fill the buffer from the stream if
-     * it is fully consumed during this operation.
-     *
-     * @param byte_type The location to store the read byte.
-     *
-     * @return bool True if the byte was successfully read and filling the byte
-     *              buffer was successful (if needed).
-     */
-    bool ReadByte(byte_type &) noexcept;
-
-    /**
-     * Read a number of bits from the byte buffer. The least-significant bits in
-     * the provided data type will be filled, starting from the position pointed
-     * to by the provided number of bits. Fill the buffer from the stream if the
-     * number of bits to read exceeds the number of bits available.
-     *
-     * @tparam DataType The data type of the location to store the read bits.
-     *
-     * @param byte_type The number of bits to read.
-     * @param DataType The location to store the read bits.
-     *
-     * @return byte_type The number of bits successfully read.
-     */
-    template <typename DataType>
-    byte_type ReadBits(byte_type, DataType &) noexcept;
-
-    /**
-     * Read a number of bits from the byte buffer without discarding those bits.
-     * The least-significant bits in the provided data type will be filled,
-     * starting from the position pointed to by the provided number of bits.
-     * Fill the buffer from the stream if the number of bits to peek exceeds the
-     * number of bits available.
-     *
-     * @tparam DataType The data type of the location to store the peeked bits.
-     *
-     * @param byte_type The number of bits to peek.
-     * @param DataType The location to store the peeked bits.
-     *
-     * @return byte_type The number of bits successfully peeked.
-     */
-    template <typename DataType>
-    byte_type PeekBits(byte_type, DataType &) noexcept;
-
-    /**
-     * Discard a number of bits from the byte buffer. Should only be used after
-     * a successful call to PeekBits.
-     *
-     * @param byte_type The number of bits to discard.
-     */
-    void DiscardBits(byte_type) noexcept;
-
-    /**
-     * Check if the stream has reached end-of-file and the byte buffer has been
-     * fully consumed.
-     *
-     * @return bool True if the stream has been fully consumed.
-     */
-    bool FullyConsumed() const noexcept;
-
-private:
-    /**
-     * Read from the stream to fill the byte buffer.
-     *
-     * @return bool True if any bits were actually read
-     */
-    bool refillBuffer() noexcept;
-
-    /**
-     * Read from the stream to fill a byte buffer.
-     *
-     * @tparam DataType The type of the byte buffer to fill.
-     *
-     * @param DataType The byte buffer to fill.
-     * @param byte_type The number of bytes to read.
-     *
-     * @return byte_type The number of bytes actually read.
-     */
-    template <typename DataType>
-    byte_type fill(DataType &, byte_type) noexcept;
-
-    std::istream &m_stream;
-    byte_type m_remainder;
-};
-
-/**
- * Traits for type safety in the BitStream template methods.
- *
- * @author Timothy Flynn (trflynn89@gmail.com)
- * @version July 7, 2019
- */
-struct BitStreamTraits
-{
-    /**
-     * Define a trait for testing if type T is an unsigned integral type.
-     */
-    template <typename T>
-    using is_unsigned_integer = std::bool_constant<
-        std::is_integral_v<std::decay_t<T>> &&
-        std::is_unsigned_v<std::decay_t<T>> &&
-        !std::is_same_v<bool, std::decay_t<T>>>;
-
-    template <typename T>
-    constexpr inline static bool is_unsigned_integer_v =
-        is_unsigned_integer<T>::value;
-
-    /**
-     * Define a trait for testing if type T is buffer_type.
-     */
-    template <typename T>
-    using is_buffer_type =
-        std::bool_constant<std::is_same_v<buffer_type, std::decay_t<T>>>;
-
-    template <typename T>
-    constexpr inline static bool is_buffer_type_v = is_buffer_type<T>::value;
-};
-
 //==============================================================================
 template <typename DataType>
 constexpr inline DataType BitStream::GenerateMask(const DataType bits)
@@ -320,143 +75,6 @@ constexpr inline DataType BitStream::GenerateMask(const DataType bits)
     constexpr auto numberOfBits = std::numeric_limits<DataType>::digits;
 
     return filledMask >> (numberOfBits - bits);
-}
-
-//==============================================================================
-template <typename DataType>
-void BitStreamWriter::WriteBits(DataType bits, byte_type size) noexcept
-{
-    static_assert(
-        BitStreamTraits::is_unsigned_integer_v<DataType>,
-        "DataType must be an unsigned integer type");
-
-    // If there are more bits to write than are available in the byte buffer,
-    // break the bits into two chunks.
-    if (size > m_position)
-    {
-        const byte_type diff = size - m_position;
-
-        // Fill the remainder of the byte buffer with as many bits as are
-        // available, and flush it onto the stream.
-        m_buffer |= (static_cast<buffer_type>(bits) >> diff);
-        flushBuffer();
-
-        // Then update the input bits to retain only those bits that have not
-        // been written yet.
-        bits &= GenerateMask<DataType>(diff);
-        size = diff;
-    }
-
-    const byte_type diff = m_position - size;
-
-    m_buffer |= static_cast<buffer_type>(bits) << diff;
-    m_position = diff;
-}
-
-//==============================================================================
-template <typename DataType>
-void BitStreamWriter::flush(const DataType &buffer, byte_type bytes) noexcept
-{
-    static_assert(
-        BitStreamTraits::is_unsigned_integer_v<DataType>,
-        "DataType must be an unsigned integer type");
-
-    const DataType data = byte_swap(buffer);
-    m_stream.write(reinterpret_cast<const std::ios::char_type *>(&data), bytes);
-}
-
-//==============================================================================
-template <typename DataType>
-byte_type BitStreamReader::ReadBits(byte_type size, DataType &bits) noexcept
-{
-    static_assert(
-        BitStreamTraits::is_unsigned_integer_v<DataType>,
-        "DataType must be an unsigned integer type");
-
-    if constexpr (BitStreamTraits::is_buffer_type_v<DataType>)
-    {
-        // See PeekBits for why 64-bit read operations must be split.
-        const byte_type sizeHigh = size / 2;
-        const byte_type sizeLow = size - sizeHigh;
-        std::uint32_t bitsHigh, bitsLow;
-
-        const byte_type bitsReadHigh = PeekBits(sizeHigh, bitsHigh);
-        DiscardBits(bitsReadHigh);
-
-        const byte_type bitsReadLow = PeekBits(sizeLow, bitsLow);
-        DiscardBits(bitsReadLow);
-
-        bits = (static_cast<DataType>(bitsHigh) << sizeLow) | bitsLow;
-        return bitsReadHigh + bitsReadLow;
-    }
-    else
-    {
-        const byte_type bitsRead = PeekBits(size, bits);
-        DiscardBits(bitsRead);
-
-        return bitsRead;
-    }
-}
-
-//==============================================================================
-template <typename DataType>
-byte_type BitStreamReader::PeekBits(byte_type size, DataType &bits) noexcept
-{
-    static_assert(
-        BitStreamTraits::is_unsigned_integer_v<DataType>,
-        "DataType must be an unsigned integer type");
-
-    // 64-bit peek operations not fully supported because the byte buffer could
-    // be in a state where it cannot be refilled. For example, consider reading
-    // 6 bits and then 64 bits. After the 6-bit read, there will be 58 bits left
-    // in the buffer - not enough for the 64 bit read. The buffer must then be
-    // refilled, but it cannot because there is less than 1 byte available.
-    //
-    // Ideally, that operation could be split into two peeks (fill the given
-    // bits with the 58 available, then refill the entire byte buffer). But this
-    // would invalidate the PeekBits/DiscardBits semantic. Would need to support
-    // putting bits back onto the stream.
-    static_assert(
-        !BitStreamTraits::is_buffer_type_v<DataType>,
-        "PeekBits only supports types smaller than buffer_type");
-
-    if ((size > m_position) && !refillBuffer())
-    {
-        return 0;
-    }
-
-    // If there are more bits to peek than are available in the byte buffer,
-    // then only the remaining bits can be returned.
-    if (size > m_position)
-    {
-        const byte_type shift = size - m_position;
-        const DataType buffer = static_cast<DataType>(m_buffer);
-
-        bits = (buffer & GenerateMask<DataType>(m_position)) << shift;
-        return m_position;
-    }
-    else
-    {
-        const byte_type shift = m_position - size;
-        const DataType buffer = static_cast<DataType>(m_buffer >> shift);
-
-        bits = buffer & GenerateMask<DataType>(size);
-        return size;
-    }
-}
-
-//==============================================================================
-template <typename DataType>
-byte_type BitStreamReader::fill(DataType &buffer, byte_type bytes) noexcept
-{
-    static_assert(
-        BitStreamTraits::is_unsigned_integer_v<DataType>,
-        "DataType must be an unsigned integer type");
-
-    m_stream.read(reinterpret_cast<std::ios::char_type *>(&buffer), bytes);
-    buffer = byte_swap(buffer);
-
-    return static_cast<byte_type>(m_stream.gcount());
 }
 
 } // namespace fly
