+++ fly/types/bit_stream/bit_stream_types.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <cstdint>
+
+namespace fly {
+
+typedef std::uint8_t byte_type;
+typedef std::uint16_t word_type;
+typedef std::uint64_t buffer_type;
+
+} // namespace fly
